backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
requests==2.32.5
pydantic==2.11.10
google-genai==0.8.0
python-multipart==0.0.12
python-dotenv==1.0.1
pymupdf==1.24.9
python-docx==1.1.2
asyncpg==0.30.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
rapidfuzz==3.9.0
